Support for USI Life Support
This is just a first attempt for feedback. It uses the smallest of the three sized tanks instead of the nice pretty parts from USI Life Support, but it doesn't break if you remove USI Life Support. It also uses the radial mounted Xenon tanks for the minipacks.

The only problem with cloning is that this introduces a hard dependency on MM that wasn't there before. I'm open to other ideas though!